pyproject.toml
[project]
name = "super-agent-platform"
version = "0.1.0"
description = "Plataforma conversacional comercial enterprise, modular y white-label."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.1,<0.111.0",
    "uvicorn[standard]>=0.29.0,<0.30.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29.0,<0.30.0",
    "alembic>=1.13.1,<2.0.0",
    "pydantic>=2",
    "pydantic-settings>=2.2.1,<3.0.0",
    "structlog>=24.1.0,<25.0.0",
    "pyyaml>=6.0.1,<7.0.0",
    "httpx>=0.27.0,<0.28.0",
    "pgvector>=0.2.5,<1.0.0",
    "apscheduler>=3.10.4,<4.0.0",
    "python-jose>=3.3.0,<4.0.0",
    "python-multipart>=0.0.9,<0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1,<9.0.0",
    "pytest-asyncio>=0.23.6,<0.24.0",
    "ruff>=0.3.5,<0.4.0",
    "black>=24.3.0,<25.0.0",
    "mypy>=1.9.0,<2.0.0",
    "types-PyYAML>=6.0.12.12,<7.0.0",
]

[tool.poetry]
name = "super-agent-platform"
version = "0.1.0"
description = "Plataforma conversacional comercial enterprise, modular y white-label."
authors = ["Super Agent Platform Team"]
readme = "README.md"
packages = [
    { include = "core" },
    { include = "adapters" },
    { include = "api" },
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
sqlalchemy = { extras = ["asyncio"], version = ">=2.0" }
asyncpg = "^0.29.0"
alembic = "^1.13.1"
pydantic = ">=2"
pydantic-settings = "^2.2.1"
structlog = "^24.1.0"
pyyaml = "^6.0.1"
httpx = "^0.27.0"
pgvector = "^0.2.5"
apscheduler = "^3.10.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
ruff = "^0.3.5"
black = "^24.3.0"
mypy = "^1.9.0"
types-PyYAML = "^6.0.12.12"

[tool.black]
line-length = 100
target-version = ["py311"]
extend-exclude = '''
(
    pytest-cache-files-.*
  | tmp.*
  | \.pytest_tmp
)
'''

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = ["core", "core.*"]
strict = true

[[tool.mypy.overrides]]
module = ["pytest.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
norecursedirs = [".pytest_tmp", "pytest-cache-files-*", "tmp*"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
